move duplicate functions to helper
